scrapy/requirements.txt
scrapy==1.4.0
python-docx==0.8.6
ped==1.5.0
html2text==2016.9.19
Pillow==4.2.1
validators=0.12.0
